reset: Add driver for dispmix reset
This is an reset driver to implement a reset controller
device DISPMIX on IMX8MM and IMX8MN platforms. Dispmix
reset is used to reset or enable related buses and clks
for the submodules in DISPMIX.

All the dispmix resets are divided into three subgroups:
sft_rstn, clk_en and mipi_rst, and each of them contains
several reset lines to control several different modules
on and off in DISPMIX which doesn't require the standard
reset flow, but only line assert and deassert operations.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>

The following patches from the NXP tree were squashed
into this patch:

6d73c7f3ae06 reset: Add driver for dispmix reset
7a139f7c2521 reset: imx8m: Correct clock name for dispmix driver
4ec8c8184529 reset: Kconfig: use 'ARCH_MXC' for reset dispmix