ARM: imx: disable cpuidle state2 by default for i.MX6SX/SL/UL/ULL
When the CPU enters idle state2 the interrupt latency increases
significantly. We can observe, when using RS485 mode of the UART,
that the switching off the RTS signal is delayed by about 2 to 3 ms.
This is caused by the TXDC interrupt being delayed and results in
collissions on the bus.

Disabling "low power idle" mode state2 fixes the behaviour and
causes the interrupt to be serviced immediately. Therefore we
disable state2 by default. It can be reenabled at runtime by

echo 0 > /sys/devices/system/cpu/cpu0/cpuidle/state2/disable

Signed-off-by: Eberhard Stoll <eberhard.stoll@kontron.de>
Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>