mtd: spi-nor: Add support for Macronix MX25R1635F
The MX25R1635F is similar to the MX25V8035F that is already
supported. The main differences are a wider supply voltage
range and twice the size (16Mb).

It was tested on the Kontron Electronics i.MX8MM SoM (N8010)
using raw read and write from and to the mtd device and
the 'flash_erase' command.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>