ARM: dts: Move watchdog from Kontron i.MX6UL/ULL board to SoM
The watchdog's WDOG_ANY signal is used to trigger a POR of the SoC,
if a soft reset is issued. As the SoM hardware connects the WDOG_ANY
and the POR signals, the watchdog node itself and the pin
configuration should be part of the common SoM devicetree.
Let's move it from the baseboard's devicetree to its proper place.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>