NFC: nxp-nci: Report wakeup events to the PM core
In order to not only wakeup from sleep when the NFC controller is set as
wakeup source, but also properly report the wakeup event, we add a call
to pm_wakeup_dev_event() in the IRQ handler. This allows us to query
the system wakeup sources and distinguish wakeup events triggered by
the NFC controller from other wakeup sources.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>