drm/bridge: tc358x67: Fix pixel clock polarity for currently used 5" display
The driver is currently not capable of using the parameters from the panel to set
things like the clock polarity. We only use this in one setup, so as a quick fix
let's just change the hardcoded output pixel clock polarity to match the requirements
of the 5" display we use.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>