regulator: pca9450: Reset PRESET_EN bit to fix BUCK1/2/3 voltage setting
The driver uses the DVS registers PCA9450_REG_BUCKxOUT_DVS0 to set the
voltage for the buck regulators 1, 2 and 3. This has no effect as the
PRESET_EN bit is set by default. This causes the preset values to be
used instead, which are set to 850 mV by default.

To fix this we reset the PRESET_EN bit at time of initialization.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>