ARM: dts: imx6ull-a1-gw: Fix the RTC description
The imx6ull-a1-gw board has a M41T00 RTC, and not an Epson RTC
as provided by imx6ul-kontron-n6x1x-s.dtsi.

This causes the following error:

rtc-rv8803 3-0032: Unable to read register 0x0e

Fix it by disabling the Epson RTC and describe the M41T00 RTC that
is populated on the board.

Signed-off-by: Fabio Estevam <festevam@gmail.com>